Starting to get ready for remote programming tools
I now know how to add programming nodes and connect them,
using only their names and the names of their outputs.
Starting to get ready for remote programming tools

I now know how to add programming nodes and connect them,
using only their names and the names of their outputs.

Everything is done through (expansive) GetType and GetField
calls. That seems to be working, so the remaining parts are
understand how to :
* Connect the impulses inputs
* Select and output when multiple ones are present
  (I'm convinced that GetField will do the trick)
* Deal with generic types, if need be.

And the last part is, of course, how to setup a nice protocol.
For the moment, this will look like a bad CSV system.

Meanwhile, I'm trying to understand how I could do the nodes
representation in Godot. I think I got one that seems to be
'kind' of working for a quick test, but I'm really wondering
if I can't integrate that to Godot VisualScripting system.

Signed-off-by: VR Voyage <avatarsetup_voyage@miouyouyou.fr>