Specific output and impulse selection code
I now know how to select a specific output and link it
to an input. And I also know how to link Impulses by names.

I still lack array inputs/outputs support but this will
do for now.

The next step is the text 'protocol' to deserialize and
serialize Logix Nodes.
I'll put a Godot project in place to represent the ndoes.

Signed-off-by: VR Voyage <avatarsetup_voyage@miouyouyou.fr>