NeosVR Remote LogiX programming system : Tested and working !
With the editor, made with Godot, on the other side.

There's still no direct support for dynamic variables definitions,
which is a big flaw, but I'll go with that for the first alpha
release.

At the moment, constants and instantation of generic types is
supported, though you'll have to provide the entire definition
of the generic type like this :
FrooxEngine.LogiX.Operators.NotNullNode<System.Single>

Still, IT WORKS ! YAY !

So I'll release the first alpha in the following days.
I still have to make the project 'CMake'-able, in order to have
autodetection of the DLL dependencies.

Signed-off-by: VR Voyage <avatarsetup_voyage@miouyouyou.fr>