Preparing for the first alpha release
A 'connect' button was added. When clicking it, it adds the
websocket client component and set it up automatically.
And clicking on 'disconnect', it removes it.

I strongly advise you to remove any remaining Websocket
client after restarting the game, since the handlers
setup is not saved, so the remaining websocket clients
won't be connected to the plugin, actually.

Still, this is completely wrorking, so wrapping up for
first release.

Signed-off-by: VR Voyage <avatarsetup_voyage@miouyouyou.fr>